+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## March 16, 2021
 
+- Release v0.3.
 - Widgets Added: ExchangeRate
 - Widgets Updated: Calendar (styles)